==> 53 - end to end deep learning project using ann/1.1 - ann - build_model_classification_problem_practical.ipynb <==
**ANN Practical > Classification problem**
***Dataset***
- Churn dataset
- Output feature
    - Exited the bank as a customer (0 or 1)
        - 0 = Person continues being a customer at the bank
        - 1 = Person stopped being a customer at teh bank.
        
- Input features
    - Non important features: RowNumber, CustomerId, Surname
    - Important features: Geography, Gender, Age, Tenure, Balance, NumOfProducts, HasCrCard, IsActiveMember,EstimatedSalary

***Problem Statement***
- Based on the Input features find out if the person entering the bank is going to leave the Bank or not.
- Exited = 1 (Leaves the bank)
- Exited = 0 (Does not leave the bank)
- Type of problem: Classification Problem (Customer leaves the blank or Customer does not leave the bank)

***Steps to Follow***
- Step 1) Load the dataset
- Step 2) Preprocessing the data
  - Step 2.1) Drop unnecessary columns
- Step 3 - Encoding categorical data
  - Step 3.1) Convert Gender from Male, Female to numeric values 0 and 1
  - Step 3.2) Convert Geography from Text to Numeric using OneHotEncoding
    - Step 3.2.1) Combine the encoded Geography columns back to the main data and drop the original Geography column
- Step 4 ) Save the LabelEncoder and OneHotEncoder using pickle for future use
- Step 5) Divide the data into features - dependent and independent features
- Step 6) Split the data into training and testing sets
- Step 7) Scaling using Standard Scalar
  - Step 7.1) Scale the features using StandardScaler
  - Step 7.2) Save the standardScalar in a pickle file for future use
- Step 8) Build the ANN model
  - Step 8.1) Create the Input Layers, Hidden Layers and Output Layer in our ANN model
  - Step 8.2) Optimizer
  - Step 8.3) Loss - BinaryCrossEntropy (here)
  - Step 8.4) Pass the Optimizer and Loss to the ann_model
  - Step 8.5) Callback > Setup the Tensorboard
    - Step 8.5.1) Create a directory for storing logs and define log file format
    - Step 8.5.2) Initialize the TensorBoard
  - Step 8.6) Callback > Setup Early Stopping
  - Step 8.7) Train the ANN model
  - Step 8.8) Save the model in H5 format
- Step 9) Bring up the Tensorboard
  - Step 9.1) Load Tensorboard Extension
  - Step 9.2) Launch the Tensorboard

    

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, LabelEncoder
import pickle
from sklearn.preprocessing import OneHotEncoder
import datetime

# ANN imports
from tensorflow.keras.models import load_model
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.callbacks import EarlyStopping,TensorBoard


# - Step 1) Load the dataset
data = pd.read_csv("./resources/data/Churn_Modelling.csv")
data.head()

# Step 2) Preprocessing the data

# Step 2.1) Drop unnecessary columns
# axis=1 (columnwise)
data = data.drop(columns=['RowNumber', 'CustomerId', 'Surname'],axis=1)
data.head()

# Step 3 - Encoding categorical data
# LabelEncoder converts Categorical data into numeric data

# Step 3.1) Convert Gender from Male, Female to numeric values 0 and 1
labelEncoder = LabelEncoder()
data['Gender']=labelEncoder.fit_transform(data['Gender'])
data.head()

# Step 3.2) Convert Geography from Text to Numeric using OneHotEncoding
# Not using LabelEncoder as assigining 1,2,3 to Germany, France, Spain may mean that Spain(3) is greater than Germany(1) and may confuse the model
# sparse_output = “sparse” refers to how the encoded output matrix is stored in memory.
# - sparse_output = False = Dense numpy array - Full matrix with all zeros and ones
# - sparse_output = True = Memory-efficient, stores only non-zero values
oneHotEncoder = OneHotEncoder(sparse_output=False)
geography_encoded = oneHotEncoder.fit_transform(data[['Geography']])
oneHotEncoder.get_feature_names_out(['Geography'])
geography_encoded = pd.DataFrame(geography_encoded, columns=oneHotEncoder.get_feature_names_out(['Geography']))
geography_encoded.head()


# Step 3.2.1) Combine the encoded Geography columns back to the main data and drop the original Geography column
data = data.drop(columns=['Geography'], axis=1)
data = pd.concat([data, geography_encoded], axis=1)
data.head()

# Step 4 ) Save the LabelEncoder and OneHotEncoder using pickle for future use
# wb = write binary mode
with open('./resources/pickle/1_label_encoder_gender.pkl', 'wb') as le_file: 
    pickle.dump(labelEncoder, le_file)  
    
with open('./resources/pickle/1_onehot_encoder_geography.pkl', 'wb') as ohe_file: 
    pickle.dump(oneHotEncoder, ohe_file)

# Step 5) Divide the data into features - dependent and independent features
# independent features = x = all columns except 'Exited' column
# dependent feature = y = 'Exited' column
x = data.drop(columns=['Exited'], axis=1)  # independent features
y = data['Exited']  # dependent feature

print('x=',x.head())
print('y=',y.head())


# Step 6) Split the data into training and testing sets
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)
print('x_train=',x_train.shape)
print('x_test=',x_test.shape)
print('y_train=',y_train.shape)
print('y_test=',y_test.shape)

# Step 7) Scaling using Standard Scalar

# Step 7.1) Scale the features using StandardScaler
# Scaling is done so that different features with different units(e.g. weights in kilograms, distance in meters) and different ranges do not bias the model
# These are scaled to a common scale with mean=0 and standard deviation=1
standardScalar = StandardScaler()
x_train = standardScalar.fit_transform(x_train) # fit_transform means learn and transform
x_test = standardScalar.transform(x_test)    # transform only means use the learned parameters (from x_train) to transform x_test
print('Scaled x_train=',x_train[:5])
print('Scaled x_test=',x_test[:5])

# Step 7.2) Save the standardScalar in a pickle file for future use
with open('./resources/pickle/1_standard_scaler_xtrain_xtest.pkl', 'wb') as scaler_file: 
    pickle.dump(standardScalar, scaler_file)    

***Artificial Neural Network (ANN) implementation***

# No of rows in x_train
print('x_train.shape[0] = ',x_train.shape[0])

# No of columns in x_train
print('x_train.shape[1] = ',x_train.shape[1])

# overall shape of x_train
print('x_train.shape = ',x_train.shape)

# Step 8) Build the ANN model
# Step 8.1) Create the Input Layers, Hidden Layers and Output Layer in our ANN model

# Input Layer and Hidden Layer 1:
# - ann_model.add(Dense(units=64, activation='relu', input_shape=(x_train.shape[1],)))
# - This statement does the following two things:-
# -1) Input Layer: input_shape=(x_train.shape[1] creates the input layer with the no of neurons equal to the no of features(columns) in x_train given by x_train.shape[1]
# -2) Hidden Layer 1: units = 64, tells how many neurons to create in that layer (Hidden Layer 1)

# Keras.Dense library is used to create Hidden Layers with the specified number of neuros
# - e.g. ann_model.add(Dense(units=32, activation='relu')) # Hidden Layer 2

# Hidden Layer 2:
# units = 32 -> create 32 neurons in Hidden Layer 2
# Activation function = relu (Others that can be used in hidden layers = Leaky Relu, TanH, etc)

# Output Layer:
# units = 1 -> create 1 output neuron giving Binary Output i.e. Exited (output feature)= 0/1 


# Activation function
# Hidden Layer - Sigmoid, Tanh, Renu, LeakyRelu, etc
# Output Layer - Sigmoid (for Binary classification (here)), Softmax(for Multi-class classification)



ann_model = Sequential()
ann_model.add(Dense(units=64, activation='relu', input_shape=(x_train.shape[1],))) # Input Layer and Hidden Layer 1
ann_model.add(Dense(units=32, activation='relu')) # Hidden Layer 2
ann_model.add(Dense(units=1, activation='sigmoid')) # Output Layer

ann_model.summary()

# Step 8.2) Optimizer
# Learning rate controls how big a step the optimizer takes when updating the weights during backward propogation
# In short the Learning rate determines how fast the model learns
# It is a constant that is initialized by us usually to a very SMALL value.

# Optimizer classes
# (1) Adadelta (2) Adafactor (3) Adagrad (4) Adam (5) AdamW (6) Adamx
# Adam in the best optimizer

# Weight updation Formula (during back propogation) = W(New) = W(Old) - (Learning Rate) * (Derivative of Loss) / (Derivative of Weight)
# - W(New) = New Weight
# - W(Old) = Old Weight
ann_optimizer=tf.keras.optimizers.Adam(learning_rate=0.01)


# Step 8.3) Loss - BinaryCrossEntropy(here)
# Loss function for Binary Classification - Use BinayCrossEntropy 
# Loss function for Multiclass Classification - Use SparseCrossEntropy
binary_cross_entropy_loss = tf.keras.losses.BinaryCrossentropy()


# Step 8.4) Pass the Optimizer and Loss to the ann_model

# Accuracy algorithms:
# Classification problems: metrics = ["accuracy"]
# Regression problems : metrics = MSE, MAE, etc
ann_model.compile(optimizer=ann_optimizer,loss=binary_cross_entropy_loss, metrics=["accuracy"])

# Step 8.5) Callback > Setup the Tensorboard
from tensorflow.keras.callbacks import EarlyStopping,TensorBoard

# Step 8.5.1) Create a director for storing logs and define log file format
# - These logs will be used by Tensorboard library for Visualization
logs_directory = "./resources/logs/1/" +datetime.datetime.now().strftime("%Y%m%d-%H%M%S")

# Step 8.5.2) Initialize the TensorBoard
tensorflow_callback = TensorBoard(log_dir=logs_directory, histogram_freq=1)


# Step 8.6) Callback > Setup Early Stopping
# patience = 10, means keep running the ANN till atleast 10 epochs even if the loss is not decreasing.
# After 10 epochs if the loss is not decreasing then apply Early Stopping and stop the ANN - Dont run it for the complete 100 epochs as show in Step 8.6

# - We can train our model for any number of epochs e.g. 100
# - In each epoch the Loss value should continuously decrease with the main goal to get close to the Global Minima
# - If the Loss value is not decreasing then we need NOT run the model for all epochs and can stop the model early (defined by patience attribute, e.g. patience = 10, meaning run for atleasat 10 epochs even if Loss is not decreasing)
# - This early stoppage of the model is known as Early Stopping. 

# restore_best_weights = True
# - “When training stops, roll back the model to the weights from the epoch where validation loss was lowest.”'
# - Keras keeps track of the best epoch (lowest val_loss)
# - This is the preferred approach

# restore_best_weights = False
# - Model keeps the weights of the last Epoch run
early_stopping_callback=EarlyStopping(monitor='val_loss',patience=10,restore_best_weights=True)




# Step 8.7) Train the ANN model
history = ann_model.fit(x_train, y_train, validation_data=(x_test, y_test), epochs=100,
callbacks=[tensorflow_callback, early_stopping_callback])

# Step 8.8) Save the model in H5 format
ann_model.save('./resources/pickle/1_ann_model.h5')

# Step 9) Bring up the Tensorboard
# Tensorboard is to visualize all your logs while training your entire model.

# Step 9.1) Load Tensorboard Extension
%load_ext tensorboard

# Step 9.2) Launch the Tensorboard
%tensorboard --logdir ./resources/logs/

# Step 10) Load the labelEncoder, onehot_encoder and standard_scaler from pickle files 
# - labelEncoder (Step 3.1) - used to convert Gender from Male, Female to numeric values 0 and 1
# - oneHotEncoder (Step 3.2) - used to convert Geography from Text to Numeric using OneHotEncoding
# - standardScalar (Step 7.2) - Scale the x_train and x_test  data  using StandardScalar so all columns have values in a similar range
# - ann_model (Step 8.8) - Trained Model using ANN (Artificial Neural Network)

with open('./resources/pickle/1_label_encoder_gender.pkl', 'rb') as le_file: 
    loaded_label_encoder_gender = pickle.load(le_file)  
with open('./resources/pickle/1_onehot_encoder_geography.pkl', 'rb') as ohe_file:               
    loaded_onehot_encoder_geography = pickle.load(ohe_file)   
with open('./resources/pickle/1_standard_scaler_xtrain_xtest.pkl', 'rb') as scaler_file: 
    loaded_scaler_xtrain_xtest = pickle.load(scaler_file)

### Load the trained model
model=load_model('./resources/pickle/1_ann_model.h5')

# Step 11) - Create new input_data for prediction. (JSON format)
# Step 11.1) Convert the input_data to a DataFrame
# Input Data is in JSON format when created (below)

input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Male',
    'Age': 40,
    'Tenure': 3,
    'Balance': 60000,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000
}

# Convert the input_data to a DataFrame
input_df = pd.DataFrame([input_data])
print('Input DataFrame:')
print(input_df.shape)


# Step 12) - Before prediction, preprocess this data:-
# Preprocessing:-
# - Convert "Geography" to numeric using OneHotEncoding
# - Scale the data using StandardScaler
# - Convert Gender to numeric using LabelEncoder


# Step 12.1) Convert "Geography" to numeric using OneHotEncoding
# Convert the Geography of input_data( i.e. France) to Numeric using One Hot Encoder
geography_ohe_encoded = loaded_onehot_encoder_geography.transform([[input_data['Geography']]])

# Convert to DataFrame
geography_ohe_encoded_df = pd.DataFrame(geography_encoded, columns=loaded_onehot_encoder_geography.get_feature_names_out(['Geography']))

# OHE Encoded Geography DataFrame
# Output: Geography France = 1.0, Geography Germany = 0.0, Geography Spain = 0.0
geography_ohe_encoded_df

# Step 12.2) Convert "Gender" in input_data to numeric using LabelEncoder
gender_label_encoded = loaded_label_encoder_gender.transform([input_data['Gender']])

# Convert the Gender label encoded to DataFrame
gender_encoded_df = pd.DataFrame(gender_label_encoded, columns=['Gender'])

# Label Encoded Gender DataFrame
# Output: Gender: 1
gender_encoded_df

# Step 12.3) Combine the encoded Geography and Gender columns back to the input_df and drop the original Geography and Gender
# Put the Gender encoded column back to input_df
input_df['Gender']=gender_encoded_df


# Drop the origional text Geography(having value France) column from input_df 
input_df = input_df.drop(columns=['Geography'], axis=1)


# Concatenate the input_df dataframe and geography numeric dataframe
input_df = pd.concat([input_df, geography_ohe_encoded_df], axis=1)

# Print
print(input_df.shape)

# Step 13) Scaling the input_df data using StandardScaler
input_df_scaled = loaded_scaler_xtrain_xtest.transform(input_df)

input_df_scaled

# Step 14) Make Prediction using the trained model - ann_model
# Step 14.1) Predict the probability that the customer will exit the bank (Exit = 1) on the input data.
predicted_value = ann_model.predict(input_df_scaled)
print('predicted_value = ',predicted_value)

prediction_probability = predicted_value[0][0]
print('predicted_value probability = ',prediction_probability) # 0.035 = 3.5% probability


# Step 14.2) Apply customer messaging based on the output
if prediction_probability > 0.5:
    print("The customer will exit the bank")
else:
    print("The customer is not likely to exit the bank")
