==> 53 - end to end deep learning project using ann/2 - ann - build_model_regression_problem_practical.ipynb <==
**ANN Practical > Regression problem**

***Dataset***
- Churn dataset
- Output feature
    - Estimated Salary
        
        
        
- Input features
    - Non important features: RowNumber, CustomerId, Surname
    - Important features: Geography, Gender, Age, Tenure, Balance, NumOfProducts, HasCrCard, IsActiveMember

***Problem Statement***
- Predict the Estimated Salary
- Type of problem: Regression problem (Predict the salaries of a group of people)


***Steps to Follow***

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, LabelEncoder, OneHotEncoder
import pickle
import datetime

# ANN related imports
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense

# Tensorflow
from tensorflow.keras.callbacks import EarlyStopping, TensorBoard


# Step 1 - Load the dataset
data=pd.read_csv('./resources/data/Churn_Modelling.csv')
data.head()

# Step 2 - Preprocess the data
# Step 2.1 -  Drop unnecessary columns
data = data.drop(columns=['RowNumber', 'CustomerId', 'Surname'])
data.head()


# Step 2.2) Encode the categorical column Gender (Convert from Male/Female to 0/1)
label_encoder_gender = LabelEncoder()
data['Gender'] = label_encoder_gender.fit_transform(data['Gender'])
data.head()

# Step 2.3) Encode the column Geography using OneHotEncoder
onehot_encoder_geography = OneHotEncoder(handle_unknown='ignore', sparse_output=False)
geography_encoded = onehot_encoder_geography.fit_transform(data[['Geography']])
geography_encoded_df = pd.DataFrame(geography_encoded, columns=onehot_encoder_geography.get_feature_names_out(['Geography']))
data = pd.concat([data.drop(columns=['Geography']), geography_encoded_df], axis=1)
data.head()

# Step 3 - Split the data into features and target
# x = inputs = all columns except EstimatedSalary
# y = output = EstimatedSalary column
x = data.drop(columns=['EstimatedSalary'],axis=1)
y = data['EstimatedSalary']      


# Step 4 - Split the data into training and testing sets
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)


# Step 5 - Scale the features using StandardScaler
standard_scaler_xtrain_xtest = StandardScaler()
x_train = standard_scaler_xtrain_xtest.fit_transform(x_train)
x_test = standard_scaler_xtrain_xtest.transform(x_test)
x_train


# Step 6 - Save the Encoders (LabelEncoder, OneHotEncoder) and Scaler (StandardScaler) using pickle
with open('./resources/pickle/2_label_encoder_gender.pkl', 'wb') as f:
     pickle.dump(label_encoder_gender, f) 

with open('./resources/pickle/2_onehot_encoder_geography.pkl', 'wb') as f:
    pickle.dump(onehot_encoder_geography, f)

with open('./resources/pickle/2_standard_scaler_xtrain_xtest.pkl', 'wb') as f:
    pickle.dump(standard_scaler_xtrain_xtest, f)

# Step 7 - Build the ANN model
# Input Layer = input_shape=(x_train.shape[1]
# Hidden Layer 1 = 64 (neurons), activation='relu'
# Hidden Layer 2 = 32 (neurons), activation='relu'
# Output Layer = 1 neuron
ann_model = Sequential([
    Dense(64, activation='relu', input_shape=(x_train.shape[1],)),
    Dense(32, activation='relu'),
    Dense(1)  # Output layer for regression
])

## Compile the model
ann_model.compile(optimizer='adam',loss='mean_absolute_error',metrics=['mae'])

ann_model.summary()

# Step 8 - Set up TensorBoard
# Step 8.1 - Write the logs to the log directory
log_dir = "./resources/logs/2/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tensorboard_callback = TensorBoard(log_dir=log_dir, histogram_freq=1)

# Step 8.2 - Set up Early Stopping to prevent overfitting
early_stopping_callback = EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)

# Step 9 - Train the ANN model created in Step 7
ann_model_trained = ann_model.fit(
    x_train, y_train,
    validation_data=(x_test, y_test),
    epochs=100,
    callbacks=[early_stopping_callback, tensorboard_callback]
)

# Step 10 - Load the TensorBoard
%load_ext tensorboard
%tensorboard --logdir ./resources/logs/2/
